requirements.txt

discord.py==2.2.2
